%!TEX root = paper.tex
\subsection{Conceptual Model}
\label{ssec:ontology}
The conceptual model used to define the web service wrappers within this application has been influenced by WSDL, the main web service definition language, and semantic approaches such as WSMO and OWL-S. It belongs to a more complex conceptual model for FAST, which can be grouped into three levels: the gadget and screenflow level at the top, the level of individual screens in the middle, and the level of web services at the bottom. 
In short, a screenflow is an executable (in a Web browser) set of screens, while a screen is a set of building blocks with a specific purpose. These building blocks --- \emph{forms} for the UI, \emph{operators} for data manipulation and \emph{back-end services} --- all share the same structure. Every building block has a set of actions (operations in WSDL), each of which needs a set of pre-conditions (inputs) to be fulfilled in order to be executed, and provides a set of post-conditions (outputs) to other building blocks. These pre- and post-conditions are defined as individual RDF triples. This means that conditions can be modelled as graph patterns using SPARQL notation \cite{sparql2008spec}, where e.g. the post-condition of a login service such as ``there exists a user'' will be transformed into a simple pattern like ``a variable \texttt{?user} is of type \texttt{sioc:User}'', or ``\texttt{?user a sioc:User}''. Using this mechanism, extended by RDFS entailment rules, services and other building blocks can be matched automatically. The publishing and discovery platform (catalogue) can employ this functionality to support the discovery of web services based on the current user needs (expressed in the same way as pre-conditions). For further details on the conceptual model, see its definition in \cite{moeller2010fast_ontology}.



git-svn-id: https://svn.forge.morfeo-project.org/fast-fp7project@982 d8b0c59b-6544-0410-94d9-83a5658a058e